+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/atomist/rug/compare/0.17.2...HEAD
 
+### Fixed
+
+-   Don't set default parameter values if they are null
+    https://github.com/atomist/rug/issues/458
+
 ## [0.17.2] - 2017-03-22
 
 [0.17.2]: https://github.com/atomist/rug/compare/0.17.1...0.17.2